chore: Refactor portfolio section and add detailed case studies
WORK IN PROGRESS - playing with the idea of a simple case study and detailed for each item. Currently put in images and the overlay needs a lot of work, and creating portfolio folder separately for simple case studies screens.

Replaced CaseStudyCard with PortfolioCard, updated image assets, and refactored CaseStudies component to Portoflio with new image imports and structure. Added detailed and simple versions of Shelter and NFCC case study pages, updated routing in App.tsx, and adjusted HomePage to use the new portfolio component. Removed the old CaseStudyCard and updated file naming for consistency.